Add createDialog method to ProfilesFragment
Centralizes layout inflation and dialog
creation. Receives a title, a message and an
optional hint to be displayed in the dialog's
EditText view; if the hint is sent as null, the
dialog won't display its EditText.

DialogCallback interface defines an
onConfirm method that receives a nullable input
text to be used as the confirmation button's
action; its return value indicates whether the
dialog should be dismissed or not after its
execution.