Add default dialog style to ProfileFragment dialogs (#123)
* Add dialog_simple_edit_text layout for dialogs

It has a title, a description for an input field,
an input field, a cancel and a OK button

* Use dialog_simple_edit_text in Add Profile

* Adjust focus highlight corners to 2dp

* Add createDialog method to ProfilesFragment

Centralizes layout inflation and dialog
creation. Receives a title, a message and an
optional hint to be displayed in the dialog's
EditText view; if the hint is sent as null, the
dialog won't display its EditText.

DialogCallback interface defines an
onConfirm method that receives a nullable input
text to be used as the confirmation button's
action; its return value indicates whether the
dialog should be dismissed or not after its
execution.

* Use createDialog in show dialog methods

showAddDialog, showSwitchDialog, showLogoutDialog
and showRemoveDialog were changed to make use of
createDialog.

* Replace deprecated call to navigate back to Home